refactor: move major2 types to global type folder + change imports to absolute for consistency
