+++ src/parse/parse.ts
@@ -1,11 +1,11 @@
 import nearly from "nearley";
 import type { ParsedCatalogEntry } from "./types";
-import { HRowType, HSectionType } from "../tokenize";
-import type { HRow, TextRow, TokenizedCatalogEntry } from "../tokenize";
+import { HRowType, HSectionType } from "@/tokenize";
+import type { HRow, TextRow, TokenizedCatalogEntry } from "@/tokenize";
 import { writeFile } from "fs/promises";
-import { FileName } from "../classify";
+import { FileName } from "@/classify";
 import grammar from "./grammar";
-import type { Major2, Section } from "@/major2";
+import type { Major2, Section } from "@/types";
 
 export const parseRows = (rows: HRow[]) => {
   const parser = new nearly.Parser(nearly.Grammar.fromCompiled(grammar));
